@@ -14,30 +14,30 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.2",
-    "@angular/animations": "^5.0.0",
-    "@angular/cdk": "^5.2.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/flex-layout": "^2.0.0-beta.12",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/material": "^5.2.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.5",
-    "@swimlane/ngx-datatable": "^11.2.0",
-    "@swimlane/ngx-dnd": "^3.1.3",
-    "angular2-chartjs": "^0.4.1",
+    "@angular/animations": "^6.0.2",
+    "@angular/cdk": "^6.0.2",
+    "@angular/common": "^6.0.2",
+    "@angular/compiler": "^6.0.2",
+    "@angular/core": "^6.0.2",
+    "@angular/flex-layout": "^5.0.0-beta.14",
+    "@angular/forms": "^6.0.2",
+    "@angular/http": "^6.0.2",
+    "@angular/material": "^6.0.2",
+    "@angular/platform-browser": "^6.0.2",
+    "@angular/platform-browser-dynamic": "^6.0.2",
+    "@angular/router": "^6.0.2",
+    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
+    "@swimlane/ngx-datatable": "^12.0.0",
+    "@swimlane/ngx-dnd": "^4.0.1",
+    "angular2-chartjs": "^0.5.1",
     "angular2-json-schema-form": "^0.7.0-alpha.1",
-    "angular2-notifications": "^0.7.4",
-    "angular2-text-mask": "^8.0.4",
+    "angular2-notifications": "^1.0.2",
+    "angular2-text-mask": "^9.0.0",
     "bootstrap": "^4.0.0-beta",
-    "c3": "^0.4.18",
+    "c3": "^0.6.1",
     "core-js": "^2.4.1",
     "css-animator": "^2.1.1",
-    "d3": "^4.12.0",
+    "d3": "^5.4.0",
     "famfamfam-flags": "^1.0.0",
     "font-awesome-scss": "^1.0.0",
     "hammerjs": "^2.0.8",
@@ -49,36 +49,38 @@
     "ng2-ui-switch": "^1.0.2",
     "ng2-validation": "^4.2.0",
     "ngx-bar-rating": "^1.1.0",
-    "ngx-bootstrap": "^2.0.2",
+    "ngx-bootstrap": "^3.0.0",
     "ngx-carousel": "^1.3.5",
     "ngx-chips": "^1.5.10",
-    "ngx-color-picker": "^5.0.4",
-    "ngx-perfect-scrollbar": "^5.0.1",
-    "rxjs": "^5.5.2",
+    "ngx-color-picker": "^6.1.0",
+    "ngx-perfect-scrollbar": "^6.0.0",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
     "screenfull": "^3.3.2",
     "sweetalert2": "^7.0.3",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
-    "@angular/cli": "1.5.5",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@types/jasmine": "~2.5.53",
+    "@angular-devkit/build-angular": "^0.6.3",
+    "@angular/cli": "^6.0.3",
+    "@angular/compiler-cli": "^6.0.2",
+    "@angular/language-service": "^6.0.2",
+    "@types/jasmine": "~2.8.7",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
+    "@types/node": "~10.1.1",
     "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "jasmine-core": "~3.1.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.2",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^2.0.0",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
+    "karma-jasmine-html-reporter": "^1.1.0",
+    "protractor": "~5.3.2",
     "standard-version": "^4.3.0",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "ts-node": "~6.0.3",
+    "tslint": "~5.10.0",
+    "typescript": "^2.7.2"
   }
 }
