@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import{ Observable } from 'rxjs/Observable';
+import{ Observable } from 'rxjs';
 import{ EsPhoto } from './es-photo';
 import { catchError, map, tap } from 'rxjs/operators';
 
