@@ -10,13 +10,28 @@ import { JsonSchemaFormService }  from 'angular2-json-schema-form';
   selector: 'q-dataTabel-widget-component',
   template: `
 
+  <form  #dataTable="ngForm" (ngSubmit)="onSubmit(dataTable)">
+    <div class="form-group">
+      <label>Name</label>
+      <input type="text" name="Name"/>
+    </div>
+    <div class="form-group">
+      <label>Code</label>
+      <input type="text" name="Code"/>
+    </div>
+    <div class="form-group">
+    <button type="submit" class="btn btn-success" [disabled]="!dataTable.form.valid">Submit</button>
+    </div>
+
+  </form>
+
 
   <ngx-datatable
     class="data-table table-responsive"
     [columnMode]="'force'"
-    [rows]="rows" [rowHeight]="60" >
+    [rows]="rows" [rowHeight]="60" sortAscending="'icon-search'">
 
-    <ngx-datatable-column [name]="col.name" *ngFor="let col of cols" [sortable]="false">
+    <ngx-datatable-column [name]="col.name" *ngFor="let col of cols">
 
       <ng-template ngx-datatable-cell-template let-rowIndex="rowIndex" let-value="value" let-row="row">
       <span
@@ -49,13 +64,14 @@ export class QDataTableWidgetComponent implements OnInit {
   @Input() layoutNode: any;
   @Input() layoutIndex: number[];
   @Input() dataIndex: number[];
+  options: any;
   constructor(private jsf: JsonSchemaFormService) { }
 
   ngOnInit() {
-
+  this.options = this.layoutNode.options || {};
   console.log(this.jsf.data);
 
-  console.log(this.layoutNode);
+  console.log('options', this.options);
 
   for(let key in this.layoutNode.options.rows.properties) {
     let obj = this.layoutNode.options.rows.properties[key];
@@ -79,6 +95,10 @@ export class QDataTableWidgetComponent implements OnInit {
      console.log('UPDATED!', this.rows[rowIndex][cell]);
 
     this.jsf.data[this.layoutNode.name] = [...this.rows];
+  }
+
+  onSubmit(formData) {
+    console.log('formData', formData);
   }
 
 
