@@ -1,9 +1,11 @@
+
+import {interval as observableInterval, Observable, Subscription} from 'rxjs';
+
+import {map} from 'rxjs/operators';
 import {Component, OnDestroy, OnInit} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
-import {Subscription} from 'rxjs/Subscription';
-import 'rxjs/add/observable/interval';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+
+
+
 
 @Component({
   selector: 'app-task-details',
@@ -30,10 +32,10 @@ export class TaskDetailsComponent implements OnInit, OnDestroy {
 
   ngOnInit() {
     this.future = new Date(this.futureString);
-    this.$counter = Observable.interval(1000).map((x) => {
+    this.$counter = observableInterval(1000).pipe(map((x) => {
       this.diff = Math.floor((this.future.getTime() - new Date().getTime()) / 1000);
       return x;
-    });
+    }));
 
     this.subscription = this.$counter.subscribe((x) => this.message = this.dhms(this.diff));
   }
